Inline NoSealedClassMocks
