--- Source/Moq.Analyzers.Test/Data/NoSealedClassMocks.Sealed.cs
@@ -1,16 +1,0 @@
-using System;
-using Moq;
-
-namespace NoSealedClassMocks.Sealed;
-
-internal sealed class FooSealed { }
-
-internal class Foo { }
-
-internal class MyUnitTests
-{
-    private void Sealed()
-    {
-        var mock = new Mock<FooSealed>();
-    }
-}
